+++ src/chat.ts
@@ -3,10 +3,11 @@ import * as vscode from "vscode";
 import cl100kBase from "tiktoken/encoders/cl100k_base.json";
 import { Tiktoken } from "tiktoken/lite";
 
-import { APP_NAME, globalStates, internalCommands } from "./configuration";
-import { Profile } from "./profiles";
-import { MetadataModelService } from "./services/MetadataModelService";
+import { APP_NAME, getConfig, globalStates, internalCommands } from "./configuration";
+import { Profile } from "./contracts/types";
 import { Disposable } from "./provider";
+import { extractCodeBlocks, getBaseUrl } from "./util";
+import { IMetadataModelService } from "./contracts/IMetadataModelService";
 
 export class ChatParticipantProvider extends Disposable {
     public _id: string = "ChatParticipantProvider";
@@ -27,7 +28,7 @@ Examples, but use the properties from the metadata in your answers:
     private participant: vscode.ChatParticipant;
     constructor(
         private context: vscode.ExtensionContext,
-        private metadataService: MetadataModelService,
+        private metadataService: IMetadataModelService,
     ) {
         super();
         this.participant = vscode.chat.createChatParticipant(
@@ -69,12 +70,12 @@ Examples, but use the properties from the metadata in your answers:
             );
             return;
         }
-        const cleanedXml = this.metadataService.getFilteredMetadataXml(text);
+        const cleanedXml = this.metadataService.getFilteredMetadataXml(text, getConfig());
         const dataModel = await this.metadataService.getModel(profile);
 
         const replacements: Record<string, string> = {
             metadata: cleanedXml,
-            base: profile.baseUrl,
+            base: getBaseUrl(profile.baseUrl),
             version: dataModel.getODataVersion(),
         };
 
@@ -115,7 +116,7 @@ Examples, but use the properties from the metadata in your answers:
         for await (const fragment of chatResponse.text) {
             stream.markdown(fragment);
             buffer.push(fragment);
-            const codeBlocks = this.extractCodeBlocks(buffer.join(""));
+            const codeBlocks = extractCodeBlocks(buffer.join(""));
             if (codeBlocks.length === 1) {
                 const query = codeBlocks[0].trim();
 
@@ -129,16 +130,4 @@ Examples, but use the properties from the metadata in your answers:
             }
         }
     };
-
-    private extractCodeBlocks(response: string): string[] {
-        const codeBlockRegex = /```odata(?:\w+)?\n([\s\S]*?)\n```/g;
-        let match;
-        const codeBlocks: string[] = [];
-
-        while ((match = codeBlockRegex.exec(response)) !== null) {
-            codeBlocks.push(match[1]); // Extracts the code inside the triple backticks
-        }
-
-        return codeBlocks;
-    }
 }