+++ src/services/QueryRunner.ts
@@ -1,12 +1,10 @@
-import { getConfig } from "../configuration";
-import { AuthKind, Profile } from "../profiles";
 import { Agent, fetch } from "undici";
+import { getConfig } from "../configuration";
+import { AuthKind, Profile } from "../contracts/types";
+import { IQueryRunner } from "../contracts/IQueryRunner";
+import { IFileReader } from "../contracts/IFileReader";
 
-export interface IFileReader {
-    readFile(path: string): Promise<Uint8Array>;
-}
-
-export class QueryRunner {
+export class QueryRunner implements IQueryRunner {
     private fileReader: IFileReader;
 
     constructor(fileReader: IFileReader) {