+++ src/test/integration/formatter.test.ts
@@ -1,21 +1,15 @@
 import * as assert from "assert";
 import * as vscode from "vscode";
-import { setupTestEnvironment } from "./util";
-
-async function setTabSize(size: number) {
-    const config = vscode.workspace.getConfiguration("editor");
-    await config.update("tabSize", size, vscode.ConfigurationTarget.Global);
-}
-
-async function setEOL(eol: string) {
-    const config = vscode.workspace.getConfiguration("files");
-    await config.update("eol", eol, vscode.ConfigurationTarget.Global);
-}
+import { setupTests } from "./testutil";
+import { ODataDocumentFormatter } from "../../formatting";
+import { SyntaxParser } from "../../parser/syntaxparser";
 
 suite("Document Formatter", () => {
-    setup(async () => {
-        await setTabSize(4);
-        await setEOL("\n");
+    let context: vscode.ExtensionContext;
+    let formatter: ODataDocumentFormatter;
+    suiteSetup(async () => {
+        ({ context } = await setupTests());
+        formatter = new ODataDocumentFormatter(new SyntaxParser());
     });
 
     const testCases = [