+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2025-04-13
+
+### Added
+
+- Allow filtering of metadata for profile
+- Copy Query command: Add current combined single-line query to clipboard
+- OData function signature help
+- Show response on query runner failure
+
+### Fixed
+
+- Register commands only for odata language
+- Property names in output from diagnostics fixed
+
+## [0.1.5] - 2025-04-06
+
+### Added
+
+- Marketplace integration improvements
+
 ## [0.1.4] - 2025-04-04
 
 ### Fixed